✨new: Colocando vue em component pages
Agora parte de formularios esta componentizada
