+++ resources/js/ComponentsPage/Admin/Company/Edit.vue
@@ -1,0 +1,5 @@
+<template>
+
+</template>
+<script setup>
+</script>
