--- resources/js/ComponentsPage/Admin/PaymentPlan/pages.js
@@ -1,4 +1,0 @@
-import PaymentPlanCreate from '@/ComponentsPage/Admin/PaymentPlan/Create.vue';
-export{
-    PaymentPlanCreate
-}
