+++ resources/js/ComponentsPage/js/company.js
@@ -1,0 +1,8 @@
+import Create from '@/ComponentsPage/Admin/Company/Create.vue';
+import Edit from '@/ComponentsPage/Admin/Company/Edit.vue';
+import List from '@/ComponentsPage/Admin/Company/List.vue';
+export{
+    Create,
+    Edit,
+    List
+}
