+++ resources/js/ComponentsPage/js/payment_plan.js
@@ -1,0 +1,8 @@
+import Create from '@/ComponentsPage/Admin/PaymentPlan/Create.vue';
+import Edit from '@/ComponentsPage/Admin/PaymentPlan/Edit.vue';
+import List from '@/ComponentsPage/Admin/PaymentPlan/List.vue';
+export{
+    Create,
+    Edit,
+    List
+}
