+++ resources/js/Pages/Admin/PaymentPlan/Create.vue
@@ -2,11 +2,11 @@
     <AppLayout title="Plano Pagamento" title_page="Novo plano de pagamento">
         <template #center>
             <Panel header="Cadastro">
-                <PaymentPlanCreate></PaymentPlanCreate>
+                <payment_plan.Create></payment_plan.Create>
             </Panel>
         </template>
     </AppLayout>
 </template>
 <script setup>
-    import { PaymentPlanCreate } from '@/ComponentsPage/Admin/PaymentPlan/pages.js';
+    import * as payment_plan from '@/ComponentsPage/js/payment_plan.js';
 </script>
